src/requirements.txt
langchain-google-vertexai>=2.0.24
fastapi==0.115.6
uvicorn==0.32.1
websockets==13.0
python-multipart==0.0.20
pydantic==2.10.4
python-dotenv==1.0.1
requests==2.32.3
aiofiles==24.1.0
jinja2==3.1.3
langgraph==0.3.16
google-auth==2.28.1
google-cloud-aiplatform==1.92.0